+++ Runtime/ReferencePickerAttribute.cs
@@ -1,5 +1,4 @@
 ﻿#if UNITY_2021_3_OR_NEWER
-using System;
 using UnityEngine;
 
 namespace SaintsField